Move BannerViewModel.java
